refactor: Translator 인터페이스를 infra 에서 domain 으로 이동
